Write preliminary robust import code
This code shows that there might be several issues with the data
both aggregated so far and to come. There are discrepancies in
the feature questions, as well as the answers, and the code table
could definitely do with an overhaul.

This code also hard-codes the documents to be imported for now,
and is naïve about their file types, so this serves only
as a proof of concept, not as THE import script for the future.